+++ platform/jvm/gradle-test-app/src/main/java/io/bitdrift/gradletestapp/init/CaptureSdkInitializer.kt
@@ -13,6 +13,7 @@ import android.content.SharedPreferences
 import android.util.Log
 import android.widget.Toast
 import com.bugsnag.android.Bugsnag
+import com.google.firebase.FirebaseApp
 import com.google.firebase.crashlytics.FirebaseCrashlytics
 import io.bitdrift.capture.Capture
 import io.bitdrift.capture.CaptureResult
@@ -67,7 +68,7 @@ object CaptureSdkInitializer {
         )
 
         return when (persistedSdkConfigResult) {
-            
+
             is PersistedSdkConfigResult.Success -> {
                 startCaptureSdk(persistedSdkConfigResult.captureSdkInitSettings, applicationContext)
                 logPreviousRunInfoToBitdrift()
@@ -100,9 +101,7 @@ object CaptureSdkInitializer {
                 is CaptureResult.Success -> {
                     val logger = startResult.value
                     Log.d("bitdrift","SDK started successfully. sessionId=${logger.sessionId}, sessionUrl=${logger.sessionUrl}")
-                    Sentry.setExtra(bitdriftSessionUrlKey, logger.sessionUrl)
-                    Bugsnag.addMetadata("bitdrift_session_url", bitdriftSessionUrlKey,  logger.sessionUrl )
-                    FirebaseCrashlytics.getInstance().setCustomKey(bitdriftSessionUrlKey, Capture.Logger.sessionUrl ?: "")
+                    addSessionUrlToThirdPartySdks(context, logger.sessionUrl)
                 }
 
                 is CaptureResult.Failure -> {
@@ -247,6 +246,24 @@ object CaptureSdkInitializer {
         }
     }
 
+    private fun addSessionUrlToThirdPartySdks(applicationContext: Context, sessionUrl: String) {
+        if (Sentry.isEnabled()) {
+            Sentry.setExtra(bitdriftSessionUrlKey, sessionUrl)
+        }
+
+        if (Bugsnag.isStarted()) {
+            val frontendTabName = "bitdrift_session_url"
+            Bugsnag.addMetadata(frontendTabName, bitdriftSessionUrlKey, sessionUrl)
+        }
+
+        FirebaseApp.getApps(applicationContext)
+            .firstOrNull { it.name == FirebaseApp.DEFAULT_APP_NAME }
+            ?.let {
+                FirebaseCrashlytics.getInstance()
+                    .setCustomKey(bitdriftSessionUrlKey, sessionUrl)
+            }
+    }
+
     private class CustomerIssueReportCallback : IssueReportCallback {
         override fun onBeforeReportSend(report: Report) {
             Capture.Logger.logInfo(
@@ -260,7 +277,6 @@ object CaptureSdkInitializer {
             ) {
                 "IssueReportCallback.onBeforeReportSend"
             }
-            FirebaseCrashlytics.getInstance().setCustomKey("bitdrift_session_id_on_before_send", report.sessionId)
         }
     }